+++ wagtail/wagtailadmin/forms.py
@@ -1,14 +1,17 @@
 from __future__ import unicode_literals
 
 import copy
+from itertools import groupby
 
 import django
 from django import forms
 from django.contrib.auth import get_user_model
 from django.contrib.auth.forms import AuthenticationForm, PasswordResetForm
+from django.contrib.auth.models import Group, Permission
 from django.core import validators
-from django.db import models
+from django.db import models, transaction
 from django.forms.widgets import TextInput
+from django.template.loader import render_to_string
 from django.utils import timezone
 from django.utils.translation import ugettext as _
 from django.utils.translation import ugettext_lazy, ungettext
@@ -16,7 +19,7 @@ from modelcluster.forms import ClusterForm, ClusterFormMetaclass
 from taggit.managers import TaggableManager
 
 from wagtail.wagtailadmin import widgets
-from wagtail.wagtailcore.models import Page, Collection
+from wagtail.wagtailcore.models import Page, Collection, GroupCollectionPermission
 
 
 class URLOrAbsolutePathValidator(validators.URLValidator):
@@ -364,3 +367,156 @@ class BaseCollectionMemberForm(forms.ModelForm):
             self.instance.collection = self.collections[0]
 
         return super(BaseCollectionMemberForm, self).save(commit=commit)
+
+
+class BaseGroupCollectionMemberPermissionFormSet(forms.BaseFormSet):
+    """
+    A base formset class for managing GroupCollectionPermissions for a
+    model with CollectionMember behaviour. Subclasses should provide attributes:
+    permission_types - a list of (codename, short_label, long_label) tuples for the permissions
+        being managed here
+    permission_queryset - a queryset of Permission objects for the above permissions
+    default_prefix - prefix to use on form fields if one is not specified in __init__
+    template = template filename
+    """
+    def __init__(self, data=None, files=None, instance=None, prefix=None):
+        if prefix is None:
+            prefix = self.default_prefix
+
+        if instance is None:
+            instance = Group()
+
+        self.instance = instance
+
+        initial_data = []
+
+        for collection, collection_permissions in groupby(
+            instance.collection_permissions.filter(
+                permission__in=self.permission_queryset,
+            ).order_by('collection'),
+            lambda cp: cp.collection
+        ):
+            initial_data.append({
+                'collection': collection,
+                'permissions': [cp.permission for cp in collection_permissions]
+            })
+
+        super(BaseGroupCollectionMemberPermissionFormSet, self).__init__(
+            data, files, initial=initial_data, prefix=prefix
+        )
+        for form in self.forms:
+            form.fields['DELETE'].widget = forms.HiddenInput()
+
+    @property
+    def empty_form(self):
+        empty_form = super(BaseGroupCollectionMemberPermissionFormSet, self).empty_form
+        empty_form.fields['DELETE'].widget = forms.HiddenInput()
+        return empty_form
+
+    def clean(self):
+        """Checks that no two forms refer to the same collection object"""
+        if any(self.errors):
+            # Don't bother validating the formset unless each form is valid on its own
+            return
+
+        collections = [
+            form.cleaned_data['collection']
+            for form in self.forms
+            if form not in self.deleted_forms
+        ]
+        if len(set(collections)) != len(collections):
+            # collections list contains duplicates
+            raise forms.ValidationError(
+                _("You cannot have multiple permission records for the same collection.")
+            )
+
+    @transaction.atomic
+    def save(self):
+        if self.instance.pk is None:
+            raise Exception(
+                "Cannot save a GroupCollectionMemberPermissionFormSet "
+                "for an unsaved group instance"
+            )
+
+        # get a set of (collection, permission) tuples for all ticked permissions
+        forms_to_save = [form for form in self.forms if form not in self.deleted_forms]
+
+        final_permission_records = set()
+        for form in forms_to_save:
+            for permission in form.cleaned_data['permissions']:
+                final_permission_records.add((form.cleaned_data['collection'], permission))
+
+        # fetch the group's existing collection permission records for this model,
+        # and from that, build a list of records to be created / deleted
+        permission_ids_to_delete = []
+        permission_records_to_keep = set()
+
+        for cp in self.instance.collection_permissions.filter(
+            permission__in=self.permission_queryset,
+        ):
+            if (cp.collection, cp.permission) in final_permission_records:
+                permission_records_to_keep.add((cp.collection, cp.permission))
+            else:
+                permission_ids_to_delete.append(cp.id)
+
+        self.instance.collection_permissions.filter(id__in=permission_ids_to_delete).delete()
+
+        permissions_to_add = final_permission_records - permission_records_to_keep
+        GroupCollectionPermission.objects.bulk_create([
+            GroupCollectionPermission(
+                group=self.instance, collection=collection, permission=permission
+            )
+            for (collection, permission) in permissions_to_add
+        ])
+
+    def as_admin_panel(self):
+        return render_to_string(
+            self.template,
+            {'formset': self},
+        )
+
+
+def collection_member_permission_formset_factory(
+    model, permission_types, template, default_prefix=None
+):
+
+    permission_queryset = Permission.objects.filter(
+        content_type__app_label=model._meta.app_label,
+        codename__in=[codename for codename, short_label, long_label in permission_types]
+    )
+
+    if default_prefix is None:
+        default_prefix = '%s_permissions' % model._meta.model_name
+
+    class CollectionMemberPermissionsForm(forms.Form):
+        """
+        For a given model with CollectionMember behaviour,
+        defines the permissions that are assigned to an entity
+        (i.e. group or user) for a specific collection
+        """
+        collection = forms.ModelChoiceField(
+            queryset=Collection.objects.all()
+        )
+        permissions = forms.ModelMultipleChoiceField(
+            queryset=permission_queryset,
+            required=False,
+            widget=forms.CheckboxSelectMultiple
+        )
+
+    GroupCollectionMemberPermissionFormSet = type(
+        'GroupCollectionMemberPermissionFormSet',
+        (BaseGroupCollectionMemberPermissionFormSet, ),
+        {
+            'permission_types': permission_types,
+            'permission_queryset': permission_queryset,
+            'default_prefix': default_prefix,
+            'template': template,
+        }
+    )
+
+    return forms.formset_factory(
+        CollectionMemberPermissionsForm,
+        formset=GroupCollectionMemberPermissionFormSet,
+        extra=0,
+        can_delete=True
+    )
