+++ wagtail/wagtailimages/templates/wagtailimages/permissions/includes/image_permissions_formset.html
@@ -1,0 +1,5 @@
+{% extends "wagtailadmin/permissions/includes/collection_member_permissions_formset.html" %}
+{% load i18n %}
+
+{% block title %}{% trans "Image permissions" %}{% endblock %}
+{% block add_button_label %}{% trans "Add an image permission" %}{% endblock %}
